refactoring: move IApplicationContext to separate file
